Uniformly round timestamps where required
A previous attempt to introduce automatic timestamp rounding based on the
ActiveRecord adaptor introduced an intermittent bug where comparison of
timestamps would not work due to comparing rounded and non-rounded values.

To mitigate these we now round *all* timestamps uniformly if required to do so,
not just certain values that we're taking as inputs for comparison.

Bumping the version to 1.1.1 to reflect this bug fix.