Remove time rounding as a breaking change
StateOfTheNation models an interval that's open at the start and closed
at the end (i.e. it includes the start and excludes the end), and
rounding timestamps doesn't allow us to strictly use it that way.

For example, ActiveSupport's `end_of_day` convenience method returns a
timestamp that includes 999,999 microseconds, but is still within the
day and should be included in the active period of an object that's made
inactive at the start of the next day.